+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Get-DotNetVersion: Updated versions to include values for 4.8.1
+
 ## [1.1.0] - 2022-12-25
 
 ### Added