+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Get-Office365IPURL, Fixed an issue where objects where not returned by the cmdlet
+
 ## [1.3.1] - 2023-01-19
 
 ### Addded