+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.2] - 2023-08-23
+
 ### Fixed
 
 - Get-Office365IPURL, Fixed an issue where objects where not returned by the cmdlet